pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Abdullah", email = "abdullarajpoot1585@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "google-auth>=2.40.2",
    "litellm>=1.71.1",
    "python-dotenv>=1.1.0",
]

[project.scripts]
chatbot = "chatbot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
